Use `Buffer.isBuffer` instead of peeking zeroith index when wrapping (#42)

+++ library/file/src/jsMain/kotlin/io/matthewnelson/kmp/file/internal/JsPlatform.kt
@@ -50,7 +50,7 @@ internal actual inline fun File.platformReadBytes(): ByteArray = try {
 
     // Max buffer size for Node.js 16 can be larger than the
     // maximum size of the ByteArray capacity
-    if (buffer.length.toLong() >= Int.MAX_VALUE.toLong()) {
+    if (buffer.length >= Int.MAX_VALUE.toLong()) {
         throw IOException("File size exceeds limit of ${Int.MAX_VALUE}")
     }
